requirements.txt
pandas
numpy
matplotlib
Pillow
scipy
scikit-learn
ipywidgets
ipyfilechooser
jupyterlab
ipympl